+++ tests/unit/DbTest.php
@@ -1780,6 +1780,24 @@ class DbTest extends TestCase
 
         $member->offset('string');
         $this->assertEquals(0, $member->getSqlPart('offset'));
+
+        $member->offset(9848519079999155811);
+        $this->assertEquals(0, $member->getSqlPart('offset'));
+    }
+
+    public function testInvalidPage()
+    {
+        $this->initFixtures();
+        $member = $this->db('member');
+
+        // @link http://php.net/manual/en/language.types.integer.php#language.types.integer.casting.from-float
+        // (984851907999915581 - 1) * 10
+        // => 9.8485190799992E+18
+        // => (int)9.8485190799992E+18
+        // => -8598224993710352384
+        // => 0
+        $member->page(984851907999915581);
+        $this->assertEquals(0, $member->getSqlPart('offset'));
     }
 
     public function testMax()