+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 Changelog
 =========
 
-## 0.9.22
+## 0.9.22 (2016-11-22)
 
-# Record: 修复page传入过大值后,offset溢出变负数的问题 #184
+* StatsD: 增加StatsD服务,用于发送统计数据到https://github.com/etsy/statsd
+* NearCache: 增加NearCache服务,可用于缓存远程数据到本地缓存中,减少远程调用
 * Memcached: 兼容memcached 3.0 getMulti方法只支持两个参数的情况
 * Env: 允许通过本地的.env.php返回值获取环境名称
 * Error: 增加autoExit选项,在命令行下运行失败时,会调用exit返回错误码,以便外部程序判断运行结果
+* Record: 修复page传入过大值后,offset溢出变负数的问题 #184
 
 ## 0.9.21 (2016-08-04)